+++ debian/bender/usr/lib/python3/dist-packages/bender/__init__.py
@@ -1,3 +1,0 @@
-# Bender — Linux Workstation Dashboard
-# Copyright (C) 2026 Chuck Talk <chuck@nordheim.online>
-# Licensed under GPLv3 or later